+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next Version (git main)
 
 - Add a progress bar for the `diff` function.
+- Fix bug where some codecs may not return the final frame.
 
 ## Version 0.1.0-beta.5